requirements.txt
pandas==2.1.4
pyyaml==6.0.1
pydantic==2.5.3
openai
requests
